fixed captcha race condition on load
This closes #242

There was a race condition where sometimes the script of the captcha component hasn't loaded on time and it was causing a runtime error.

This commit modifies the code so that all the window methods are set first, and then the script tag is imported, this way we know that the required environment for the captcha is correctly set up.