Switching to WS provider (#275)
Fixes #223

Tested using this method:
https://polkadot.js.org/docs/api/examples/promise/upgrade-chain,
changing provider from HTTP to WS actually seems to do the trick.

The original plan was to add runtime upgrade to e2e tests, however
supporting that test setup seems to be a bit too much. It basically
requires an additional zombienet setup, which would also be different
for local development and CI.
Basically, that would be a whole new test pack with its own setup,
specifically for an issue in of `@polkadot/api`.
Still can be done, but seems to be a bit too much.