Removing ui image and docker-compose files
1. UI image.
The problem with that one, is that it's not being used. We deploy UI to
GitHub pages currently, and even if we potentially will switch from it,
deploying static site as a docker image is far from a good solution with
much higher operation costs and downtimes on deployment.
Also, fixes #279

2. docker-compose.yml's
These were made for the local development by DevOps, however, they
aren't fit for the local development (the don't provide watch mode,
re-building is painful, debug...), and they are actually fit for
running it locally if you aren't developing the app.