Using PAPI for e2e tests
* At first, e2e types were generated on postinstall and prebuild steps,
but that required metadata to be available in Docker image, even though
it's not technically used in the application code.
Making e2e types generation a manual step together with excluding tests
code from Docker image solved it better.
* .scale metadata files from Zombienet hosts saved in the codebase,
in order to be able to compile the code separately from running Zombienet
* Upgraded both Zombienet and Polkadot in the CI, as well as the docs
for e2e tests.
* Used persistent path for zombienet nodes' logs, and printing them in
CI in case of a failure.