Updated actions to latest and added retention period
We currently don't have an artifact retention period, so it defaults to 90 days.

While the artifact for the site is not too big, having one build per push and PR can quickly accumulate the amount and make us hit our hard limit.

By adding a limit to the retention day of 5 days we ensure that we get rid of the old artifacts faster than waiting for two whole months.

I also updated all the actions version to latest release