Fixing "Invalid" error
The source of the error was, that I treated the observables the wrong
way.
I expected `submitAndWatch()` to return a chain of events, to which I
could subscribe from a bunch of sources, while it actually led to side
effects being multiplied in the sources.
`shareReplay()` returns a multicast observable that actually behaves the
way I expected it to.