Form can be valid regardless of order filling out (#28)
## Goal
This adjusts the faucet form validation to allow the submit button to be enabled regardless of the order of clicking captcha or filling in an address. Address validation isn't performed until `focusout` fires. A little trickier to fix than at first blush.

Also did a yarn update to fix a browserslist warning. Thanks to @JoeCap08055 for the nudge to turn this into an issue.

Closes #27